v0.8 Sprint 1 (QA-010): decompose CommunityPersistenceEngine — close GodClass + TooManyMethods suppressions (#119)
Extracts two responsibility-bundles from CommunityPersistenceEngine into
dedicated Community-internal helpers (package-private, eu.exeris.kernel.
community.persistence):

* CommunityPersistenceMigrationRunner — owns the "read SQL migration
  resources from the classpath, split into statements on top-level
  semicolons outside single-quoted literals, strip line comments,
  execute every statement inside one transaction with explicit
  auto-commit lifecycle" workflow. 8 methods moved out (maybeRunMigrations,
  runMigrationsInTransaction, restoreAutoCommit, executeMigrationScript,
  readMigrationResource, splitSqlStatements, addStatement,
  stripLineComments). Behavioural semantics preserved verbatim against
  the V0.5.0__create_outbox.sql and V0.7.0__create_saga_state.sql scripts
  pinned by CommunityPersistenceEngineMigrationTest (2/2 green).

* CommunityPersistencePoolWarmup — owns the "hold N connections at once
  to force HikariCP to materialise them concurrently before the first
  request arrives" workflow with fail-fast on Connection.isValid().
  4 methods moved out (prewarmSharedPool, acquireWarmupConnections,
  validateWarmupConnection, releaseConnectionsQuietly).

Engine shrinks 565 → 403 LOC (−162). PMD class-level suppression block:

    Before: {PMD.CyclomaticComplexity, PMD.TooManyMethods,
             PMD.CouplingBetweenObjects, PMD.GodClass}
    After:  {PMD.CyclomaticComplexity, PMD.CouplingBetweenObjects}

The remaining two suppressions are documented in the class header with
their root cause: the per-request connection-acquisition decision tree
(request-scope vs physical, tenant pool selection, interceptor chain)
dominates remaining cyclomatic complexity; coupling reflects the SPI
integration surface (PersistenceEngine + PhysicalConnectionSource +
ConnectionInterceptor + StorageContext) which is load-bearing by
contract, not by accident. PMD.GodClass + PMD.TooManyMethods are gone.

QA-010 exit criterion (one GodClass closed per QA-* PR) met. Cumulative
v0.8 suppression count: −2 (GodClass + TooManyMethods on engine);
remaining nine GodClass-flagged classes queued under QA-011..018.

Local validation: full non-integration test sweep across all 10 modules
— BUILD SUCCESS. CommunityPersistenceEngineMigrationTest (the regression
test for the extracted migration workflow) passes against H2.
